python-service/requirements.txt
fastapi
uvicorn
jinja2
boto3
pydantic
langchain
langchain-core
langchain-google-genai
langgraph
python-dotenv
